requirements.txt
cffi==1.17.1
cryptography==44.0.2
mysql-connector-python==9.2.0
numpy==2.0.2
pandas==2.2.3
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
SQLAlchemy==2.0.39
typing_extensions==4.12.2
tzdata==2025.1
